+++ packages/orq-rc/src/orq_ai_sdk/models/conversationresponse.py
@@ -104,7 +104,7 @@ class ConversationResponse(BaseModel):
     r"""Unix timestamp in milliseconds when the conversation was last modified."""
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "conv_01kht0v9xenpgat2x942mq37zj"
+        "conv_01khtjfjtt3s2azz161c8364a5"
     )
     r"""Unique conversation identifier with `conv_` prefix."""