+++ packages/orq-rc/src/orq_ai_sdk/models/listdatasourcesop.py
@@ -148,7 +148,7 @@ class ListDatasourcesData(BaseModel):
     r"""The number of chunks in the datasource"""
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "01KHT0VAEDDPEBXNN3C5NT2HQT"
+        "01KHTJFKBPHWRGNZCP8SXY6PDF"
     )
     r"""The unique identifier of the data source"""