+++ packages/orq-rc/src/orq_ai_sdk/models/reasoningpart.py
@@ -37,7 +37,7 @@ class ReasoningPart(BaseModel):
     r"""The reasoning or thought process behind the response. Used for chain-of-thought or extended thinking."""
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "reasoning_01kht0v9hfh7pfqj1q164mwjq7"
+        "reasoning_01khtjfjf5nv7974x66spewg71"
     )
     r"""Unique identifier for the part. Format: reasoning_{ulid} (e.g., reasoning_01hxyz...)"""