+++ packages/orq-rc/src/orq_ai_sdk/models/updateconversationop.py
@@ -231,7 +231,7 @@ class UpdateConversationResponseBody(BaseModel):
     r"""Unix timestamp in milliseconds when the conversation was last modified."""
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "conv_01kht0v9y4wssgp2e8jq5rve0e"
+        "conv_01khtjfjvfrjz0edz0weaxybq3"
     )
     r"""Unique conversation identifier with `conv_` prefix."""