+++ packages/orq-rc/src/orq_ai_sdk/models/updateidentityop.py
@@ -194,7 +194,7 @@ class UpdateIdentityResponseBody(BaseModel):
     created: Optional[datetime] = None
     r"""The date and time the resource was created"""
 
-    updated: Optional[datetime] = parse_datetime("2026-02-19T03:59:23.423Z")
+    updated: Optional[datetime] = parse_datetime("2026-02-19T09:07:33.562Z")
     r"""The date and time the resource was last updated"""
 
     @model_serializer(mode="wrap")